Refactor move user-related types to api module
