+++ client/src/composables/hashedUserId.ts
@@ -2,7 +2,8 @@ import { useLocalStorage } from "@vueuse/core";
 import { storeToRefs } from "pinia";
 import { computed, type Ref, ref, watch } from "vue";
 
-import { GenericUser, useUserStore } from "@/stores/userStore";
+import type { GenericUser } from "@/api";
+import { useUserStore } from "@/stores/userStore";
 
 async function hash32(value: string): Promise<string> {
     const valueUtf8 = new TextEncoder().encode(value);
